backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.10.4
pydantic-settings==2.7.1
anthropic==0.42.0
httpx==0.28.1
python-multipart==0.0.20
websockets==14.1
pytest==8.3.4
pytest-asyncio==0.25.0
